PullRequest: graalpython/3730

+++ scripts/wheelbuilder/build_wheels.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023, 2024, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2023, 2025, Oracle and/or its affiliates. All rights reserved.
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 #
 # The Universal Permissive License (UPL), Version 1.0
@@ -50,6 +50,7 @@ The steps are:
 import hashlib
 import importlib
 import os
+import platform
 import re
 import shlex
 import shutil
@@ -65,17 +66,20 @@ from tempfile import TemporaryDirectory
 from urllib.request import urlretrieve
 
 
-def ensure_installed(name):
+def ensure_installed(name, *extra):
     try:
         return importlib.import_module(name)
     except ImportError:
-        subprocess.check_call([sys.executable, "-m", "pip", "install", name])
+        subprocess.check_call([sys.executable, "-m", "pip", "install", name, *extra])
         return importlib.import_module(name)
 
 
 def download(url, out):
-    print("Downloading", url, flush=True)
-    urlretrieve(url, out)
+    if not os.path.exists(out):
+        print("Downloading", url, flush=True)
+        urlretrieve(url, out)
+    else:
+        print("Using previously downloaded", out, flush=True)
 
 
 def extract(archive):
@@ -100,6 +104,11 @@ def create_venv():
     subprocess.check_call([binary, "-m", "venv", "graalpy"])
     print("Installing wheel with", pip, flush=True)
     subprocess.check_call([pip, "install", "wheel"])
+    if sys.platform == "win32":
+        print("Installing python-patch-ng to provide patch.exe", flush=True)
+        p = subprocess.run([pip, "install", "git+https://github.com/timfel/python-patch-ng.git"])
+        if p.returncode != 0:
+            print("Installing python-patch-ng failed, assuming patch.exe is on PATH", flush=True)
     return pip
 
 
@@ -121,66 +130,108 @@ def build_wheels(pip):
         available_scripts = {s.lower(): s for s in os.listdir(scriptdir)}
     else:
         available_scripts = {}
-    for spec in packages_to_build:
-        name, version = spec.split("==")
-        whl_count = len(glob("*.whl"))
-        script = f"{name}.{version}.{script_ext}".lower()
-        if script not in available_scripts:
-            script = f"{name}.{script_ext}".lower()
-        if script in available_scripts:
-            script = join(scriptdir, available_scripts[script])
-            env = os.environ.copy()
-            env["PATH"] = abspath(dirname(pip)) + os.pathsep + env["PATH"]
-            env["VIRTUAL_ENV"] = abspath(dirname(dirname(pip)))
-            print("Building", name, version, "with", script, flush=True)
-            if sys.platform == "win32":
-                cmd = [script, version]  # Python's subprocess.py does the quoting we need
-            else:
-                cmd = f"{shlex.quote(script)} {version}"
-            subprocess.check_call(cmd, shell=True, env=env)
-            if not len(glob("*.whl")) > whl_count:
-                print("Building wheel for", name, version, "after", script, "did not", flush=True)
-                subprocess.check_call([pip, "wheel", spec])
-        else:
+    remaining_packages = 0
+    while remaining_packages != len(packages_to_build):
+        remaining_packages = len(packages_to_build)
+        for spec in packages_to_build.copy():
+            name, version = spec.split("==")
+            whl_count = len(glob("*.whl"))
+            script = f"{name}.{version}.{script_ext}".lower()
+            if script not in available_scripts:
+                script = f"{name}.{script_ext}".lower()
+            if script in available_scripts:
+                script = join(scriptdir, available_scripts[script])
+                env = os.environ.copy()
+                env["PATH"] = abspath(dirname(pip)) + os.pathsep + env["PATH"]
+                env["VIRTUAL_ENV"] = abspath(dirname(dirname(pip)))
+                print("Building", name, version, "with", script, flush=True)
+                if sys.platform == "win32":
+                    cmd = [script, version]  # Python's subprocess.py does the quoting we need
+                else:
+                    cmd = f"{os.environ.get('SHELL', '/bin/sh')} {shlex.quote(script)} {version}"
+                p = subprocess.run(cmd, shell=True, env=env)
+                if p.returncode != 0:
+                    continue
+                if len(glob("*.whl")) > whl_count:
+                    packages_to_build.remove(spec)
+                    continue
+                print(script, "did not build a wheel, we will do so now", flush=True)
             print("Building", name, version, flush=True)
-            subprocess.check_call([pip, "wheel", spec])
+            p = subprocess.run([pip, "wheel", spec])
+            if p.returncode == 0:
+                packages_to_build.remove(spec)
+    if packages_to_build:
+        print("Failed to build all packages, the following packages failed")
+        print(packages_to_build)
+        return False
+    else:
+        return True
 
 
 def repair_wheels():
-    if sys.platform == "win32":
-        ensure_installed("delvewheel")
-        env = os.environ.copy()
-        env["PYTHONUTF8"] = "1"
-        subprocess.check_call(
-            [
-                sys.executable,
-                "-m",
-                "delvewheel",
-                "repair",
-                "-v",
-                "--exclude",
-                "python-native.dll",
-                "-w",
-                "wheelhouse",
-                *glob("*.whl"),
-            ],
-            env=env,
-        )
-    elif sys.platform == "linux":
-        ensure_installed("auditwheel")
-        subprocess.check_call(
-            [join(dirname(sys.executable), "auditwheel"), "repair", "-w", "wheelhouse", *glob("*.whl")]
-        )
-    elif sys.platform == "darwin":
-        ensure_installed("delocate")
-        subprocess.check_call(
-            [join(dirname(sys.executable), "delocate-wheel"), "-v", "-w", "wheelhouse", *glob("*.whl")]
-        )
+    whls = glob("*graalpy*.whl")
+    env = os.environ.copy()
+    env["PYTHONUTF8"] = "1"
+    env["PATH"] = abspath(dirname(sys.executable)) + os.pathsep + env["PATH"]
+    for whl in whls:
+        if sys.platform == "win32":
+            ensure_installed("delvewheel")
+            p = subprocess.run(
+                [
+                    sys.executable,
+                    "-m",
+                    "delvewheel",
+                    "repair",
+                    "-v",
+                    "--exclude",
+                    "python-native.dll",
+                    "-w",
+                    "wheelhouse",
+                    whl,
+                ],
+                env=env,
+            )
+        elif sys.platform == "linux":
+            ensure_installed("auditwheel", "patchelf")
+            p = subprocess.run(
+                [
+                    join(dirname(sys.executable), "auditwheel"),
+                    "repair",
+                    "--plat",
+                    "manylinux_2_28_x86_64" if platform.processor() == "x86_64" else "manylinux_2_28_aarch64",
+                    "-w",
+                    "wheelhouse",
+                    whl,
+                ],
+                env=env,
+            )
+        elif sys.platform == "darwin":
+            ensure_installed("delocate")
+            p = subprocess.run(
+                [
+                    join(dirname(sys.executable), "delocate-wheel"),
+                    "-v",
+                    "--ignore-missing-dependencies",
+                    "--require-archs",
+                    "arm64" if platform.processor() == "arm" else "x86_64",
+                    "-w",
+                    "wheelhouse",
+                    whl,
+                ],
+                env=env,
+            )
+        if p.returncode != 0:
+            print("Repairing", whl, "failed, copying as is.")
+            try:
+                shutil.copy(whl, "wheelhouse")
+            except:
+                pass
 
 
 if __name__ == "__main__":
     parser = ArgumentParser()
     parser.add_argument("graalpy_url")
+    parser.add_argument("--ignore-failures", action="store_true", default=False)
     args = parser.parse_args()
     ext = splitext(args.graalpy_url)[1]
     outpath = f"graalpy{ext}"
@@ -188,5 +239,7 @@ if __name__ == "__main__":
     download(args.graalpy_url, outpath)
     extract(outpath)
     pip = create_venv()
-    build_wheels(pip)
+    success = build_wheels(pip)
     repair_wheels()
+    if not success and not args.ignore_failures:
+        sys.exit(1)
